+++ src/EvoAlgo/Chromosome.cpp
@@ -4,15 +4,15 @@
 
 #include "EvoAlgo/Chromosome.h"
 
-constexpr double Chromosome::BODY_VERTEX_MAX;
-constexpr double Chromosome::WHEEL_SIZE_MAX;
+constexpr float Chromosome::BODY_VERTEX_MAX;
+constexpr float Chromosome::WHEEL_SIZE_MAX;
 
 Chromosome::Chromosome() : bodyVertices{VERTICES_NUMBER}, wheelVertices{WHEELS_NUMBER}, wheelSizes{WHEELS_NUMBER}
 {
 }
 
-Chromosome::Chromosome(std::vector<std::pair<double, double>>& bodyVertices, std::vector<int>& wheelVertices,
-                       std::vector<double>& wheelSizes) : Chromosome{}
+Chromosome::Chromosome(std::vector<std::pair<float, float>>& bodyVertices, std::vector<int>& wheelVertices,
+                       std::vector<float>& wheelSizes) : Chromosome{}
 {
     if(bodyVertices.size() != VERTICES_NUMBER)
     {
@@ -29,7 +29,7 @@ Chromosome::Chromosome(std::vector<std::pair<double, double>>& bodyVertices, std
     this->wheelSizes = wheelSizes;
 }
 
-const std::vector<std::pair<double, double>>& Chromosome::getBodyVertices() const
+const std::vector<std::pair<float, float>>& Chromosome::getBodyVertices() const
 {
     return bodyVertices;
 }
@@ -39,7 +39,7 @@ const std::vector<int>& Chromosome::getWheelVertices() const
     return wheelVertices;
 }
 
-const std::vector<double>& Chromosome::getWheelSizes() const
+const std::vector<float>& Chromosome::getWheelSizes() const
 {
     return wheelSizes;
 }
